Clearly separate the tree.
